refactor: Separate theme src into dimensions
