+++ packages/theme-data/src/baseTheme/components/checkbox.js
