+++ packages/theme-data/src/baseTheme/components/icon.js
