+++ packages/theme-data/src/baseTheme/components/label.js
