+++ packages/theme-data/src/baseTheme/unresolvedRoles.js
@@ -1,6 +1,6 @@
-import mapKeys from "../../utils/mapKeys";
-import extendTheme from "../../utils/extendTheme";
-import basics from "../../basics";
+import mapKeys from "../utils/mapKeys";
+import extendTheme from "../utils/extendTheme";
+import basics from "../basics";
 import system from "./system";
 import checkbox from "./components/checkbox";
 import formField from "./components/formField";
