+++ packages/theme-data/src/colorSchemes/darkBlue/components/label.js
