+++ packages/theme-data/src/colorSchemes/darkBlue/components/skeletonItem.js
