+++ packages/theme-data/src/colorSchemes/darkBlue/system/colorScheme.js
