+++ packages/theme-data/src/colorSchemes/darkBlue/unresolvedRoles.js
@@ -1,8 +1,8 @@
 import extendTheme from "../../utils/extendTheme";
 import mapKeys from "../../utils/mapKeys";
 import system from "./system";
-import baseTheme from "../baseTheme";
-import mediumDensityTheme from "../mediumDensityTheme";
+import baseTheme from "../../baseTheme";
+import mediumDensityTheme from "../../densities/mediumDensity";
 import formField from "./components/formField";
 import label from "./components/label";
 import skeletonItem from "./components/skeletonItem";
