+++ packages/theme-data/src/colorSchemes/lightGray/components/label.js
