+++ packages/theme-data/src/colorSchemes/lightGray/components/skeletonItem.js
