+++ packages/theme-data/src/colorSchemes/lightGray/metadata.js
