+++ packages/theme-data/src/colorSchemes/lightGray/system/colorScheme.js
