+++ packages/theme-data/src/colorSchemes/lightGray/system/index.js
