+++ packages/theme-data/src/colorSchemes/lightGray/unresolvedRoles.js
@@ -1,7 +1,7 @@
 import extendTheme from "../../utils/extendTheme";
 import mapKeys from "../../utils/mapKeys";
-import baseTheme from "../baseTheme";
-import mediumDensityTheme from "../mediumDensityTheme";
+import baseTheme from "../../baseTheme";
+import mediumDensityTheme from "../../densities/mediumDensity";
 import system from "./system";
 import formField from "./components/formField";
 import label from "./components/label";
