+++ packages/theme-data/src/colorSchemes/webLight/index.js
@@ -1,9 +1,7 @@
 import metadata from "./metadata";
 import unresolvedRoles from "./unresolvedRoles";
-import resolvedRoles from "./resolvedRoles";
 
 export default {
   ...metadata,
-  resolvedRoles,
   unresolvedRoles
 };
