+++ packages/theme-data/src/densities/mediumDensity/index.js
@@ -1,9 +1,7 @@
-import resolvedRoles from "./resolvedRoles";
 import unresolvedRoles from "./unresolvedRoles";
 import metadata from "./metadata";
 
 export default {
   ...metadata,
-  resolvedRoles,
   unresolvedRoles
 };
