+++ packages/theme-data/src/densities/mediumDensity/system/index.js
