--- packages/theme-data/src/themes/darkBlueTheme/resolvedRoles.js
@@ -1,4 +1,0 @@
-import themeConfig from "./unresolvedRoles";
-import resolveTheme from "../../utils/resolveTheme";
-
-export default resolveTheme(themeConfig);
