--- packages/theme-data/src/themes/lightGrayTheme/resolvedRoles.js
@@ -1,4 +1,0 @@
-import unresolvedRoles from "./unresolvedRoles";
-import resolveTheme from "../../utils/resolveTheme";
-
-export default resolveTheme(unresolvedRoles);
