--- packages/theme-data/src/themes/webLightTheme/resolvedRoles.js
@@ -1,4 +1,0 @@
-import resolveTheme from "../../utils/resolveTheme";
-import unresolvedRoles from "./unresolvedRoles";
-
-export default resolveTheme(unresolvedRoles);
